lstm_word_gen: a few improvements
1) tokenize our corpus differently: the regexp now used probably correctly
identifies most words and some punctuation marks. this brings the number of
unique tokens down to around 6k, which is a huge improvement to the size of
our one-hot token representations.

2) some variables were renamed to be more specific.

3) trade some cpu for a lot of RAM. previously we were filling all of our
data into a single tensor and allowing keras to batch it for us. in this
change, we generate new batches before each call to keras' Model.fit()
method. notably, keras is now batching at a batch size of 32, however, we
still feed it 128 records at a time.

4) demonstrations of text generation do not happen 1:1 with calls to
Model.fit() anymore